requirements.txt
ffmpeg
python-dotenv
PyNaCl
https://github.com/yt-dlp/yt-dlp/archive/master.tar.gz
discord.py
